File and Directory setup script
You can ignore this if you'd like but I think it's better than posting a list of commands to run and if more files or directories are needed we can just update the script rather than having to update the wiki and hope people notice.

As we are recommending people run apache in least privilleged mode. By design the server cannot create files and directories. By creating the needed files and log directory with appropriate privilleges the server can write to the necessary files